Avoid importing anything from tests/integration or KAT in tests/unit
This involves moving some code from python/tests/utils.py to
python/tests/integration/utils.py.
