+++ python/tests/integration/test_watt_scaling.py
@@ -3,7 +3,7 @@ import sys
 
 import pytest
 
-from tests.utils import install_ambassador, get_code_with_retry
+from tests.integration.utils import install_ambassador, get_code_with_retry
 from tests.kubeutils import apply_kube_artifacts, delete_kube_artifacts
 from tests.runutils import run_with_retry, run_and_assert
 from tests.manifests import qotm_manifests
